requirements.txt
APScheduler==3.10.4
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
et-xmlfile==1.1.0
filelock==3.9.0
Flask==3.0.2
Flask-Hashing==1.1
fpdf==1.7.2
fsspec==2023.4.0
idna==3.6
inflect==7.2.1
itsdangerous==2.1.2
Jinja2==3.1.3
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
mysql-connector-python==8.3.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.3
openai-whisper==20231117
openpyxl==3.1.2
pandas==2.2.2
pillow==10.2.0
protobuf==4.21.12
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sympy==1.12
tiktoken==0.6.0
torch==2.2.1+cu118
torchaudio==2.2.1+cu118
torchvision==0.17.1+cu118
tqdm==4.66.2
typeguard==4.2.1
typing_extensions==4.12.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
Werkzeug==3.0.1
